+++ src/main/java/com/sucy/skill/listener/ExperienceListener.java
@@ -2,7 +2,6 @@ package com.sucy.skill.listener;
 
 import com.sucy.skill.SkillAPI;
 import com.sucy.skill.api.enums.ExpSource;
-import com.sucy.skill.api.event.PlayerExperienceGainEvent;
 import com.sucy.skill.api.player.PlayerClass;
 import com.sucy.skill.api.player.PlayerData;
 import mc.promcteam.engine.mccore.config.CommentedConfig;
@@ -15,7 +14,8 @@ import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.inventory.CraftItemEvent;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
 
 /**
  * ProSkillAPI © 2023
@@ -24,7 +24,6 @@ import java.util.*;
 public class ExperienceListener extends SkillAPIListener {
 
     private static final String            CONFIG_KEY = "unnatural";
-    private static       Map<UUID, Double> exempt     = new HashMap<>();
 
     boolean         track;
     HashSet<String> unnatural = new HashSet<>();
@@ -35,10 +34,6 @@ public class ExperienceListener extends SkillAPIListener {
             CommentedConfig data = SkillAPI.getConfig("data/placed");
             unnatural = new HashSet<>(data.getConfig().getList(CONFIG_KEY));
         }
-    }
-
-    public static void addExemptExperience(Player player, double amount) {
-        exempt.put(player.getUniqueId(), amount);
     }
 
     @Override
@@ -88,17 +83,6 @@ public class ExperienceListener extends SkillAPIListener {
             if (yield > 0) {
                 playerClass.giveExp(yield, ExpSource.CRAFT);
             }
-        }
-    }
-
-    @EventHandler(priority = EventPriority.LOWEST)
-    public void onExperienceGain(PlayerExperienceGainEvent event) {
-        Player player = event.getPlayerData().getPlayer();
-        if (event.isCancelled()
-                && event.getSource() == ExpSource.PLUGIN
-                && exempt.containsKey(player.getUniqueId())
-                && exempt.get(player.getUniqueId()) == event.getExp()) {
-            event.setCancelled(false);
         }
     }
